requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=2.5
PyJWT>=2.0
Werkzeug>=2.0
